requirements.txt
certifi==2020.6.20
chardet==3.0.4
cryptos==1.36
ecdsa==0.16.0
idna==2.10
python-bitcoinlib==0.11.0
requests==2.24.0
six==1.15.0
urllib3==1.25.11
